Found the reason why hypothesis did not work.
There are certain in-between value, which, due to some floating point
mambojambo, results in a weaker peak near the frequency component.
for example, when f0=220.087760925293, the peak around f0 is lower than
when it is 220.0, so the height parameter of peak detection, which
determines the minimum height of a peak, turns out to be higher than
the only peak.